+++ breeze-core/src/main/kotlin/extensions/CoreExtensions.kt
@@ -93,6 +93,8 @@ import com.windea.breezeframework.core.annotations.*
 import java.lang.reflect.*
 import kotlin.reflect.*
 
+//通过这个方法可以得到泛型参数的信息
+
 /**得到指定类型的带有泛型参数信息的Java类型对象。*/
 inline fun <reified T> javaTypeOf(): Type = object : TypeReference<T>() {}.type
 
@@ -102,6 +104,7 @@ internal abstract class TypeReference<T> {
 	val type: Type = (javaClass.genericSuperclass as ParameterizedType).actualTypeArguments[0]
 }
 
+//无法直接通过方法的引用得到参数，也无法得到局部变量的任何信息
 
 /**得到指定类型的名字。*/
 @TrickApi
@@ -110,12 +113,11 @@ inline fun <reified T> nameOf(): String? = T::class.java.simpleName
 /**得到指定项的名字。适用于：类引用、属性引用、方法引用、实例。不适用于：类型参数，参数，局部变量。*/
 @TrickApi
 @JvmSynthetic
-inline fun nameOf(target: Any?): String? = when {
-	//无法直接通过方法的引用得到参数，也无法得到局部变量的任何信息
-	target == null -> null
-	target is Class<*> -> target.simpleName
-	target is KClass<*> -> target.simpleName
-	target is KCallable<*> -> target.name
-	target is KParameter -> target.name
+inline fun nameOf(target: Any?): String? = when(target) {
+	null -> null
+	is Class<*> -> target.simpleName
+	is KClass<*> -> target.simpleName
+	is KCallable<*> -> target.name
+	is KParameter -> target.name
 	else -> target::class.java.simpleName
 }