+++ breeze-core/src/test/kotlin/extensions/AnyExtensionsTest.kt
@@ -84,22 +84,20 @@
  * Breeze is blowing ...
  **********************************************************************************************************************/
 
-package com.windea.breezeframework.core.domain
+package com.windea.breezeframework.core.extensions
 
-import java.io.Serializable
-import java.time.temporal.*
+import org.junit.*
 
-/**
- * 可审计的对象。
- *
- * @property createdTime 创建时间。
- * @property lastUpdatedTime 最后更新时间。
- * @property createdBy 创建者。
- * @property lastUpdatedBy 最后更新者。
- */
-interface Auditable<T: Temporal,U: Serializable>:Serializable{
-	val createdTime:T
-	var lastUpdatedTime:T
-	val createdBy:U
-	var lastUpdatedBy:U
+class AnyExtensionsTest{
+	@Test
+	fun testConvert(){
+		"123".convert<Int>().andPrintln()
+		"123".safeConvert<Int>().andPrintln()
+	}
+
+	@Test
+	fun testCast(){
+		(123 as Number).javaClass.andPrintln()
+		(123 as Number).javaClass.andPrintln()
+	}
 }